# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PowerHub — multi-tenant hotel power/relay automation control system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties the current user can access
      responses:
        "200":
          description: Properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProperty
      tags: [properties]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProperty
      tags: [properties]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /process-types:
    get:
      operationId: listProcessTypes
      tags: [process-types]
      parameters:
        - $ref: "#/components/parameters/PropertyIdQuery"
      responses:
        "200":
          description: Process types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcessType"
    post:
      operationId: createProcessType
      tags: [process-types]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessType"
        "400":
          $ref: "#/components/responses/BadRequest"
  /process-types/{id}:
    patch:
      operationId: updateProcessType
      tags: [process-types]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessTypeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessType"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProcessType
      tags: [process-types]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
  /roles/{id}:
    patch:
      operationId: updateRole
      tags: [roles]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteRole
      tags: [roles]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PropertyIdQuery:
      name: propertyId
      in: query
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Property:
      type: object
      required: [id, name, code, currency, tariffPerKwh, timezone, active, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        currency: { type: string }
        tariffPerKwh: { type: number }
        timezone: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PropertyInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        address: { type: string }
        city: { type: string }
        pincode: { type: string }
        email: { type: string }
        phone: { type: string }
        currency: { type: string }
        tariffPerKwh: { type: number, minimum: 0 }
        timezone: { type: string }
        active: { type: boolean }
    PropertyUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        address: { type: string }
        city: { type: string }
        pincode: { type: string }
        email: { type: string }
        phone: { type: string }
        currency: { type: string }
        tariffPerKwh: { type: number, minimum: 0 }
        timezone: { type: string }
        active: { type: boolean }
...
    ProcessType:
      type: object
      required: [id, propertyId, name, cutoffMinutes, isAuto, active]
      properties:
        id: { type: integer }
        propertyId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        cutoffMinutes: { type: integer }
        isAuto: { type: boolean }
        active: { type: boolean }
    ProcessTypeInput:
      type: object
      required: [propertyId, name, cutoffMinutes, isAuto]
      properties:
        propertyId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string }
        cutoffMinutes: { type: integer, minimum: 0 }
        isAuto: { type: boolean }
        active: { type: boolean }
    ProcessTypeUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        cutoffMinutes: { type: integer, minimum: 0 }
        isAuto: { type: boolean }
        active: { type: boolean }
    Role:
      type: object
      required: [id, name, permissions, isSystem]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        permissions:
          type: array
          items: { type: string }
        isSystem: { type: boolean }
    RoleInput:
      type: object
      required: [name, permissions]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        permissions:
          type: array
          items: { type: string }
    RoleUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        permissions:
          type: array
          items: { type: string }
...