# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * PowerHub — multi-tenant hotel power/relay automation control system
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  code?: string;
  address?: string;
  city?: string;
  pincode?: string;
  email?: string;
  phone?: string;
  currency?: string;
  /** @minimum 0 */
  tariffPerKwh?: number;
  timezone?: string;
  active?: boolean;
}